Bubble Sort, etc
Lots of new methods in the ListObject file, and a 'sort' button that
initilizes the active sort on the list. No built-out way to change the
active sort yet, but you just need to change the active sort string.
A sort step is only called when all child objects have reached their
destinations.

Built out bubble sort. Seems to work.